renamed correlation.corrfunc to correlation.corrfuncs to avoid sphinx name confusion
